Track instance of Output.
Virtually all callback interfaces allow user-defined data to be passed to
the callback function via a void pointer. The Wayland library has chosen
to be unique and instead passes its own data structures as void*, which
you must cast to a proper type, such as wl_output*.

This well-thought-out design requires that the Output class has to be a
singleton in order for the callback to be able to do actual work.

Bonus points to the libwayland authors for making the callback interface
look like it actually does what you would normally expect!