Allow changing unordered_flat_{map,set} to STL for debug purposes.
The robin hood variant of the map and set has no debug support in the
tools such as debuggers or IDEs. However, the standard implementation
of the map and set would typically have support for introspection of
its contents.

The implementation details of both variants obviously is completely
different, but the interfaces are generally compatible.