Change the way Gpu devices are handled.
All VlkDevices will be now created on startup, instead of on demand.

Present support is ignored during device creation / queue setup, as
there's no surface to check present support. The assumption is that
present will always be available on the graphics queue. Data available
at the moment shows that this is always true.

An alternative solution would be to create a VlkDevice with all the
queues enabled, which would allow to properly select the present setup
during VlkSwapchain creation later on.

This change breaks backend and rendering support.