Change -s parameter from run-single-threaded to synchronize-logs.
Running the server single-threaded is not practical. The only thing this
setting changed previously was number of worker threads, and this worked
fine. However, not all threads are worker threads. Making the code logic
diverge between single- and multi-threaded options makes things harder
to read and doesn't serve any real purpose.

The intended meaning of this option was to make logs look coherent. Now
this is explicitly stated.

This change only handles the logistcs, but the log synchronization has
not yet been implemented.