Rewrite mcore rendering (part 1).
WaylandBackendWindow is removed in favor of WaylandWindow. Output
tracking is not available now, but it is not critical to have.

WaylandConnector is largely redundant, as WaylandWindow does swapchain
management now. It still needs to be incorporated into the framework,
as Connector is a parameter to the Renderable Render function.

GpuDevice now holds VlkDevice as a shared ptr, to make it work with
other interfaces.

BackendWayland no longer inherits from WaylandDisplay. If output
tracking is wanted, it should be implemented as part of WaylandDisplay.

BackendWayland render function has hardcoded render size. It must be
handled properly when WaylandWindow will be handling window sizes.