Track idle rendering state in WaylandWindow.
The listener's OnRender() method is supposed to tell if it has presented
a frame. If a frame is presented, the frame callback will be fired by the
compositor as expected.

In case OnRender() does not submit a frame, the frame callback will not
fire. This, in turn, makes it not possible to react to various events, for
example resizing the window. These window events are buffered in staging
variables, and they only take effect when the frame is about to be rendered.

To make this works, a wl_surface_commit() call was executed, to submit a
no-op change that actually made the frame callback fire. This resulted in
constant communication with the compositor, which was not optimal, as neither
the application nor the compositor were able to sleep properly. The CPU load
was especially visible on the compositor side.

With this change, WaylandWindow is now tracking if the rendering is idle
(if the frame callback will not fire). Each event that may change window
properties must now call ResumeIfIdle(), which will make a no-op surface
commit if needed, and this will in turn trigger the frame callback to fire
and resume rendering.