Initial image view handling logic.
The 'f' key resizes the window to min(bounds, image size), and makes the
image fit the window, as before. Handled by FitToExtent().

The '1' key is a new mode of operation that sets the zoom level to make
the pixels 1:1. Handled by FitPixelPerfect().

Both FitToExtent() and FitPixelPerfect() require the render extent param,
because external code may want to resize the window and then set the image
zoom level. Since the render extent of the window is staged until next
redraw, it can't be reliably retrieved from the window itself.

Resizing the window will now change how much of the image is seen, instead
of changing the image scale.

Changing DPI scale now also changes the zoom level, so that the image view
doesn't visibly change when moving the window across monitors.