Fix drag and drop edge case.
The nominal drag and drop call stack is:

1. WaylandSeat::WaylandDataEnter()
2. some dispatch
3. WaylandSeat::AcceptDndMime().

The dispatch code may want to inspect data contents to decide how to proceed
by calling WaylandSeat::GetDnd(). This function does a roundtrip to initiate
data transmission on the other end of the pipe.

The roundtrip gives the wayland server an opportunity to send some messages
that may interfere with the desired flow. In the edge case that this commit
fixes the server called WaylandSeat::DataLeave(), which cleared m_dndOffer,
making the dereference in AcceptDndMime() a segmentation fault.