Set window size after unmaximizing it.
The usual way of setting the initial window size is calling SetDevice(),
which sets the Vulkan device to be used for rendering and creating a
swapchain. Then, and only then, a Resize() call can be made to create
a swapchain. Finally, when something is rendered and a present call is
made, the Vulkan surface becomes visible to Wayland compositor and the
surface size becomes window size.

The window size is then received via a XdgToplevelConfigure() callback in
the application. This will be the size that has just been set, so there
is nothing to do.

Things are different if the user would resize the window. In such case,
the window size will be staged in m_staged, and there is a need to redraw
and present the Vulkan surface. This is achieved by staging the new window
size in m_staged and resuming the rendering via ResumeIfIdle().

The actual resizing (via swapchain recreation) is done in Update(), which
compares m_staged size against the current surface size stored in m_extent.
Note that window "resize" events that do not change the window size (like
the confirmation of the initial window size) are no-op, even if they are
not special cased.

Wayland compositors will also send an initial XdgToplevelConfigure() event
with window size set to 0x0, as an indication that the compositor does not
know the expected size of the window and expects one to be provided. This
event is ignored, as valid WaylandWindow API usage requires specifying the
size, albeit at a later stage, when Vulkan device is set.

The XdgToplevelConfigure() callback also detects window states such as
maximization or fullscreen mode. These states require that m_extent is set
in Update() to follow the expected window size stored in m_staged. At the
same time, the application may want to save the unmaximized window size at
exit, so it needs to be preserverd. This size is stored in m_floatingExtent
variable (the floating part refers to a floating window, not floating point
values).

This commit fixes the behavior when maximized state is set immediately
after the window is created.

When an initial window size is passed via the Resize() call WaylandWindow
will use it to initialize m_extent, m_staged, and m_floatingExtent and will
create a swapchain. This size is not known to the Wayland compositor until
the Vulkan surface becomes presented.

When the window is maximized right after the Resize() call, the compositor
responds via XdgToplevelConfigure() to tell the window is now maximized and
what size it should be. This sets m_staged, which then sets m_extent, but
it does not touch m_floatingExtent.

When the window is unmaximized, the compositor does not know what size it
should set for the window, as the initial window size was never seen by it.
The compositor will issue a XdgToplevelConfigure() call with 0x0 size to
make the application provide the expected size.

When this happens, and it's known that the window was previously maximized,
the event is no longer ignored. The value of m_floatingExtent at this point
has the unmaximized window size, so it can be assigned to m_staged, and
the rendering can be resumed, so that new Vulkan surface is presented.